progress on home page
